requirements.txt
Pillow
google-generativeai
python-dotenv